fix: remove test_admin.rs
